Corrections from PR
