+++ .travis.yml
@@ -1,7 +1,7 @@
 language: swift
 osx_image: xcode7.2
 before_install:
-- gem install bundler
+- sudo gem install bundler
 script:
 - REPO_SLUG="$TRAVIS_REPO_SLUG" PULL_REQUEST="$TRAVIS_PULL_REQUEST" FORCE_CARTHAGE_VERSION=true
   script/cibuild
