+++ Authentication/LogIn/Model/Email.swift
@@ -12,7 +12,10 @@ import Foundation
 public struct Email {
     
     public static func isValidEmail(raw: String) -> Bool {
-        return raw.containsString("@")
+        let emailRegEx = "^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$"
+        
+        let emailTest = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
+        return emailTest.evaluateWithObject(raw)
     }
     
     let raw: String
