+++ AuthenticationTests/LogIn/Mocks/MockSessionService.swift
@@ -1,5 +1,5 @@
 //
-//  OneMyUserSessionService.swift
+//  MockSessionService.swift
 //  Authentication
 //
 //  Created by Daniela Riesgo on 3/14/16.
@@ -12,7 +12,7 @@ import Authentication
 import ReactiveCocoa
 import enum Result.NoError
 
-final class OneMyUserSessionService: SessionServiceType {
+final class MockSessionService: SessionServiceType {
     
     private let _possibleUser: MyUser
     private let (_currentUser, _currentUserObserver) = Signal<MyUser, NoError>.pipe()
@@ -29,22 +29,15 @@ final class OneMyUserSessionService: SessionServiceType {
     }
     
     func logIn(email: Email, _ password: String) -> SignalProducer<MyUser, SessionServiceError> {
-        if email == self._possibleUser.email {
-            if password == self._possibleUser.password {
-                return SignalProducer(value: self._possibleUser).on(completed: { [unowned self] in
-                    self._eventsObserver.sendNext(.LogIn(self._possibleUser))
-                    self._currentUserObserver.sendNext(self._possibleUser)
-                    })
-            } else {
-                return SignalProducer(error: SessionServiceError.WrongPassword).on(failed: { [unowned self] _ in
-                    self._eventsObserver.sendNext(.LogInError(.WrongPassword))
-                    })
-                
-            }
+        if email == self._possibleUser.email && password == self._possibleUser.password {
+            return SignalProducer(value: self._possibleUser).on(completed: { [unowned self] in
+                self._eventsObserver.sendNext(.LogIn(self._possibleUser))
+                self._currentUserObserver.sendNext(self._possibleUser)
+            })
         } else {
-            return SignalProducer(error: SessionServiceError.InexistentUser).on(failed: { [unowned self] _ in
-                self._eventsObserver.sendNext(.LogInError(.InexistentUser))
-                })
+            return SignalProducer(error: SessionServiceError.InvalidCredentials(.None)).on(failed: { [unowned self] _ in
+                self._eventsObserver.sendNext(.LogInError(.InvalidCredentials(.None)))
+            })
             
         }
     }
